subclase basetest, added new test case to force binary
